fix(config): allow notes_dir to replace entire directory
